Added default configuration target
